third_party/blink/web_tests/permissionclient/storage-permission-detached-expected.txt

Detaching a frame should bypass it's cached permissionclient state in StorageAreaProxy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS storage.length threw exception SecurityError: Failed to read the 'length' property from 'Storage': Access is denied for this document..
PASS storage.getItem('foo') threw exception SecurityError: Failed to execute 'getItem' on 'Storage': Access is denied for this document..
PASS storage.setItem('foo', 'bar') threw exception SecurityError: Failed to execute 'setItem' on 'Storage': Access is denied for this document..
PASS storage.key(0) threw exception SecurityError: Failed to execute 'key' on 'Storage': Access is denied for this document..
PASS storage.clear() threw exception SecurityError: Failed to execute 'clear' on 'Storage': Access is denied for this document..
PASS successfullyParsed is true

TEST COMPLETE
